Commands extend LivewireFileManipulationCommand
